requirements.txt
Django==1.8
Markdown==2.6.7
django-filter==0.15.3
django-queryset-csv==1.0.1
djangorestframework==3.5.0
unicodecsv==0.14.1
